test: keep duplicate evidence coverage focused
